+++ domain/src/test/java/com/parkingcomestrue/common/domain/parking/service/ParkingFilteringServiceTest.java
@@ -11,7 +11,6 @@ import com.parkingcomestrue.common.domain.parking.Parking;
 import com.parkingcomestrue.common.domain.parking.ParkingFeeCalculator;
 import com.parkingcomestrue.common.domain.parking.ParkingType;
 import com.parkingcomestrue.common.domain.parking.PayType;
-import com.parkingcomestrue.common.domain.parking.SearchingCondition;
 import com.parkingcomestrue.common.domain.parking.TimeUnit;
 import com.parkingcomestrue.common.domain.searchcondition.FeeType;
 import java.time.LocalDateTime;
